refactor(workspaces): use shared test helpers and split acc tests
Address review feedback after rebasing on main (which added the workspace_member
data source and shared workspacetest.go helpers).

- Drop redundant `parallel = true` from `run` block in workspace_member.tftest.hcl
  (the top-level `test { parallel = true }` already covers it)
- Add `UserID` field to the package-shared `workspaceMemberAPIResponse` (defined
  in workspace_member_data_source.go) and remove the duplicate type from the
  resource file
- Switch unit tests to `package workspaces` and use the shared `newTestAdminClient`
  helper from `workspacetest.go` instead of a local duplicate
- Move acceptance tests to a separate `_acc_test.go` file in `package workspaces_test`
  to avoid the `acctest` → `provider` → `workspaces` import cycle
- Fix template note that incorrectly said billing rejection only applies on creation
  (the validator runs on update too)
- Document the new `anthropic_workspace_member` resource and the workspaces test
  helpers / package-split convention in CLAUDE.md
